outsourcing of code in libraries
